refactor: merge operations as commands (PR5, back-end refactor)
Lift the four merge writes (Book/Work/Author/Edition) out of their
Web services into BookTracker.Application command handlers, mirroring
the PR1–4 pattern: intent commands + thin IDispatcher, signatures
unchanged so the merge pages/dialogs are untouched.

- MergeBooks / MergeEditions (Application/Books), MergeWorks
  (Application/Works), MergeAuthors (Application/Authors) — each an
  ICommand<TResult> whose handler holds the verbatim transactional
  merge body; the result records move to Application (C3).
- Author compatibility rule extracted to AuthorMergeCompatibility so
  the merge-preview read and the write handler stay in lock-step
  rather than each inlining the canonical-resolution rule.
- The four *MergeService classes shrink to read-only preview loaders
  (LoadAsync), per the arc's write-only adoption — the reads stay put
  until the PR6 read-model relocation. VMs now inject IDispatcher and
  dispatch the command.
- Tests: merge write tests move to *MergeHandlerTests under
  Application/<Feature>/; the *MergeServiceTests keep the LoadAsync
  reads; VM tests mock IDispatcher.Send. 45 integration + 23 unit
  merge tests green; full solution builds clean.

Tech-debt recon (pre-PR5): confirmed TD-15 (shared find-or-create /
TagResolver) does NOT fold into the merge handlers — merges union/
reassign existing rows by id/FK and never find-or-create a lookup row.
