+++ BookTracker.Tests/Services/AuthorMergeServiceTests.cs
@@ -3,14 +3,16 @@ using BookTracker.Web.Services;
 
 namespace BookTracker.Tests.Services;
 
+// Loader-only since PR5 — the merge write moved to MergeAuthorsHandler
+// (AuthorMergeHandlerTests). These cover the merge-preview reads, including the
+// canonical-compatibility signal (shared with the handler via
+// AuthorMergeCompatibility) that the preview surfaces.
 [Trait("Category", TestCategories.Integration)]
 public class AuthorMergeServiceTests
 {
     private readonly TestDbContextFactory _factory = new();
 
     private AuthorMergeService CreateService() => new(_factory);
-
-    // ─── LoadAsync ────────────────────────────────────────────────────
 
     [Fact]
     public async Task LoadAsync_returns_both_details_with_counts_and_samples()
@@ -63,223 +65,6 @@ public class AuthorMergeServiceTests
 
         Assert.Null(result.IncompatibilityReason);
     }
-
-    // ─── MergeAsync — reassignments ───────────────────────────────────
-
-    [Fact]
-    public async Task MergeAsync_reassigns_works_and_deletes_loser()
-    {
-        var ids = await SeedAuthorsWithWorksAsync(
-            ("Douglas Preston", ["Title A", "Title B"]),
-            ("Doug Preston", ["Title C"]));
-
-        var result = await CreateService().MergeAsync(winnerId: ids[0], loserId: ids[1]);
-
-        Assert.True(result.Success);
-        Assert.Equal(1, result.WorksReassigned);
-        Assert.Equal("Douglas Preston", result.WinnerName);
-        Assert.Equal("Doug Preston", result.LoserName);
-
-        using var db = _factory.CreateDbContext();
-        Assert.Equal(3, db.Works.Count(w => w.WorkAuthors.Any(wa => wa.AuthorId == ids[0])));
-        Assert.Null(db.Authors.FirstOrDefault(a => a.Id == ids[1]));
-    }
-
-    [Fact]
-    public async Task MergeAsync_reassigns_aliases_of_loser_to_winner()
-    {
-        using var db = _factory.CreateDbContext();
-        var winner = new Author { Name = "Stephen King" };
-        var loser = new Author { Name = "S. King" };
-        db.Authors.AddRange(winner, loser);
-        await db.SaveChangesAsync();
-        // An existing pen name pointing at the loser — should reattach to winner.
-        var bachman = new Author { Name = "Richard Bachman", CanonicalAuthorId = loser.Id };
-        db.Authors.Add(bachman);
-        await db.SaveChangesAsync();
-
-        var result = await CreateService().MergeAsync(winner.Id, loser.Id);
-
-        Assert.True(result.Success);
-        Assert.Equal(1, result.AliasesReassigned);
-
-        using var verify = _factory.CreateDbContext();
-        var reloaded = verify.Authors.First(a => a.Id == bachman.Id);
-        Assert.Equal(winner.Id, reloaded.CanonicalAuthorId);
-    }
-
-    [Fact]
-    public async Task MergeAsync_promotes_winner_when_winner_was_alias_of_loser()
-    {
-        // Case: winner is "Stephen King" (an alias of loser "Steve King").
-        // After merge, winner must be promoted to canonical — otherwise its
-        // CanonicalAuthorId dangles at the deleted loser.
-        using var db = _factory.CreateDbContext();
-        var loser = new Author { Name = "Steve King" };
-        db.Authors.Add(loser);
-        await db.SaveChangesAsync();
-        var winner = new Author { Name = "Stephen King", CanonicalAuthorId = loser.Id };
-        db.Authors.Add(winner);
-        await db.SaveChangesAsync();
-
-        var result = await CreateService().MergeAsync(winner.Id, loser.Id);
-
-        Assert.True(result.Success);
-        Assert.True(result.WinnerPromotedToCanonical);
-
-        using var verify = _factory.CreateDbContext();
-        var reloaded = verify.Authors.First(a => a.Id == winner.Id);
-        Assert.Null(reloaded.CanonicalAuthorId);
-    }
-
-    [Fact]
-    public async Task MergeAsync_removes_ignored_duplicate_rows_referencing_loser()
-    {
-        var ids = await SeedAuthorsWithWorksAsync(
-            ("Douglas Preston", ["Title A"]),
-            ("Doug Preston", ["Title B"]),
-            ("Third Person", []));
-
-        using (var db = _factory.CreateDbContext())
-        {
-            db.IgnoredDuplicates.Add(new IgnoredDuplicate
-            {
-                EntityType = DuplicateEntityType.Author,
-                LowerId = Math.Min(ids[0], ids[1]),
-                HigherId = Math.Max(ids[0], ids[1])
-            });
-            // Unrelated pair that shouldn't be deleted.
-            db.IgnoredDuplicates.Add(new IgnoredDuplicate
-            {
-                EntityType = DuplicateEntityType.Author,
-                LowerId = Math.Min(ids[0], ids[2]),
-                HigherId = Math.Max(ids[0], ids[2])
-            });
-            await db.SaveChangesAsync();
-        }
-
-        await CreateService().MergeAsync(winnerId: ids[0], loserId: ids[1]);
-
-        using var verify = _factory.CreateDbContext();
-        Assert.Single(verify.IgnoredDuplicates);
-    }
-
-    // ─── MergeAsync — refusals ────────────────────────────────────────
-
-    [Fact]
-    public async Task MergeAsync_rejects_self_merge()
-    {
-        var ids = await SeedAuthorsWithWorksAsync(("Douglas Preston", []));
-
-        var result = await CreateService().MergeAsync(ids[0], ids[0]);
-
-        Assert.False(result.Success);
-        Assert.Contains("same", result.ErrorMessage!, StringComparison.OrdinalIgnoreCase);
-    }
-
-    [Fact]
-    public async Task MergeAsync_rejects_missing_winner()
-    {
-        var ids = await SeedAuthorsWithWorksAsync(("Douglas Preston", []));
-
-        var result = await CreateService().MergeAsync(winnerId: 9999, loserId: ids[0]);
-
-        Assert.False(result.Success);
-    }
-
-    [Fact]
-    public async Task MergeAsync_rejects_incompatible_canonicals()
-    {
-        using var db = _factory.CreateDbContext();
-        var z1 = new Author { Name = "Canonical A" };
-        var z2 = new Author { Name = "Canonical B" };
-        db.Authors.AddRange(z1, z2);
-        await db.SaveChangesAsync();
-        var d1 = new Author { Name = "Doug Preston", CanonicalAuthorId = z1.Id };
-        var d2 = new Author { Name = "Douglas Preston", CanonicalAuthorId = z2.Id };
-        db.Authors.AddRange(d1, d2);
-        await db.SaveChangesAsync();
-
-        var result = await CreateService().MergeAsync(d1.Id, d2.Id);
-
-        Assert.False(result.Success);
-        Assert.NotNull(result.ErrorMessage);
-    }
-
-    // ─── MergeAsync — Role-aware (Phase G) ────────────────────────────
-
-    [Fact]
-    public async Task MergeAsync_preserves_loser_role_when_winner_has_different_role_on_same_work()
-    {
-        // Winner is credited as Author on Work X; loser is credited as
-        // Translator on the same Work. Post-merge, winner must hold BOTH
-        // roles on Work X — the loser's Translator credit should NOT be
-        // dropped just because winner exists on that Work.
-        using var db = _factory.CreateDbContext();
-        var winner = new Author { Name = "Winner" };
-        var loser = new Author { Name = "Loser" };
-        db.Authors.AddRange(winner, loser);
-        await db.SaveChangesAsync();
-
-        var work = new Work
-        {
-            Title = "Work X",
-            WorkAuthors =
-            [
-                new WorkAuthor { Author = winner, Order = 0, Role = AuthorRole.Author },
-                new WorkAuthor { Author = loser,  Order = 0, Role = AuthorRole.Translator },
-            ],
-        };
-        db.Books.Add(new Book { Title = "Book X", Works = [work] });
-        await db.SaveChangesAsync();
-
-        var result = await CreateService().MergeAsync(winner.Id, loser.Id);
-
-        Assert.True(result.Success);
-
-        using var verify = _factory.CreateDbContext();
-        var rows = verify.WorkAuthors.Where(wa => wa.WorkId == work.Id).ToList();
-        Assert.Equal(2, rows.Count);
-        Assert.Contains(rows, r => r.AuthorId == winner.Id && r.Role == AuthorRole.Author);
-        Assert.Contains(rows, r => r.AuthorId == winner.Id && r.Role == AuthorRole.Translator);
-    }
-
-    [Fact]
-    public async Task MergeAsync_dedupes_loser_role_when_winner_already_has_same_role()
-    {
-        // Both winner and loser are credited as Author on Work X. Post-merge,
-        // winner has exactly one Author row on Work X (no duplicate composite-
-        // PK violation, no doubled-up credit).
-        using var db = _factory.CreateDbContext();
-        var winner = new Author { Name = "Winner" };
-        var loser = new Author { Name = "Loser" };
-        db.Authors.AddRange(winner, loser);
-        await db.SaveChangesAsync();
-
-        var work = new Work
-        {
-            Title = "Work X",
-            WorkAuthors =
-            [
-                new WorkAuthor { Author = winner, Order = 0, Role = AuthorRole.Author },
-                new WorkAuthor { Author = loser,  Order = 1, Role = AuthorRole.Author },
-            ],
-        };
-        db.Books.Add(new Book { Title = "Book X", Works = [work] });
-        await db.SaveChangesAsync();
-
-        var result = await CreateService().MergeAsync(winner.Id, loser.Id);
-
-        Assert.True(result.Success);
-
-        using var verify = _factory.CreateDbContext();
-        var rows = verify.WorkAuthors.Where(wa => wa.WorkId == work.Id).ToList();
-        Assert.Single(rows);
-        Assert.Equal(winner.Id, rows[0].AuthorId);
-        Assert.Equal(AuthorRole.Author, rows[0].Role);
-    }
-
-    // ─── Helpers ──────────────────────────────────────────────────────
 
     private async Task<List<int>> SeedAuthorsWithWorksAsync(params (string Name, string[] WorkTitles)[] data)
     {
