+++ BookTracker.Tests/ViewModels/WorkMergeViewModelTests.cs
@@ -1,3 +1,5 @@
+using BookTracker.Application;
+using BookTracker.Application.Works;
 using BookTracker.Web.Services;
 using BookTracker.Web.ViewModels;
 using NSubstitute;
@@ -8,8 +10,9 @@ namespace BookTracker.Tests.ViewModels;
 public class WorkMergeViewModelTests
 {
     private readonly IWorkMergeService _merger = Substitute.For<IWorkMergeService>();
+    private readonly IDispatcher _dispatcher = Substitute.For<IDispatcher>();
 
-    private WorkMergeViewModel CreateVm() => new(_merger);
+    private WorkMergeViewModel CreateVm() => new(_merger, _dispatcher);
 
     private static WorkMergeDetail Detail(int id, string title) =>
         new(id, title, null, "A", null, null, null, [], 0, [], null);
@@ -42,11 +45,11 @@ public class WorkMergeViewModelTests
     }
 
     [Fact]
-    public async Task MergeAsync_delegates_to_service_and_reports_success()
+    public async Task MergeAsync_dispatches_merge_command_and_reports_success()
     {
         _merger.LoadAsync(1, 2, Arg.Any<CancellationToken>())
             .Returns(new WorkMergeLoadResult(Detail(1, "A"), Detail(2, "B"), null, 0));
-        _merger.MergeAsync(1, 2, Arg.Any<CancellationToken>())
+        _dispatcher.Send(Arg.Any<MergeWorks>(), Arg.Any<CancellationToken>())
             .Returns(new WorkMergeResult(true, null, 3, 0, 0, "A", "B"));
 
         var vm = CreateVm();
@@ -57,7 +60,9 @@ public class WorkMergeViewModelTests
 
         Assert.NotNull(result);
         Assert.True(result!.Success);
-        await _merger.Received(1).MergeAsync(1, 2, Arg.Any<CancellationToken>());
+        await _dispatcher.Received(1).Send(
+            Arg.Is<MergeWorks>(c => c.WinnerId == 1 && c.LoserId == 2),
+            Arg.Any<CancellationToken>());
     }
 
     [Fact]
@@ -65,7 +70,7 @@ public class WorkMergeViewModelTests
     {
         _merger.LoadAsync(1, 2, Arg.Any<CancellationToken>())
             .Returns(new WorkMergeLoadResult(Detail(1, "A"), Detail(2, "B"), null, 0));
-        _merger.MergeAsync(1, 2, Arg.Any<CancellationToken>())
+        _dispatcher.Send(Arg.Any<MergeWorks>(), Arg.Any<CancellationToken>())
             .Returns(new WorkMergeResult(false, "boom", 0, 0, 0, null, null));
 
         var vm = CreateVm();
@@ -89,6 +94,6 @@ public class WorkMergeViewModelTests
         var result = await vm.MergeAsync();
 
         Assert.Null(result);
-        await _merger.DidNotReceiveWithAnyArgs().MergeAsync(default, default, default);
+        await _dispatcher.DidNotReceiveWithAnyArgs().Send(Arg.Any<MergeWorks>(), Arg.Any<CancellationToken>());
     }
 }
